requirements.txt
Flask
scapy
PyYAML
plotly
psutil
matplotlib
